+++ src/CargoShip.java
@@ -4,7 +4,7 @@ public class CargoShip extends WaterVehicle implements FuelConsumable, CargoCarr
     private double currentCargo;
     boolean maintenanceNeeded;
 
-    CargoShip(String id, String model, double maxSpeed, double currentMileage, boolean hasSail, int currentCargo, boolean maintenanceNeeded){
+    CargoShip(String id, String model, double maxSpeed, double currentMileage, boolean hasSail, double currentCargo, boolean maintenanceNeeded){
         super(id, model, maxSpeed, currentMileage, hasSail);
         this.currentCargo= currentCargo;
         this.maintenanceNeeded= maintenanceNeeded;
